pyproject.toml
[tool.poetry]
name = "lau_app_back"
version = "1.0.0"
description = "Módulo para Gestão Financeira"
authors = ["Renato Carnauba <contato@renatocarnauba.com.br>"]
license = "Private"
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
fastapi = "^0.92.0"
asyncpg = "^0.27.0"
dropbox = "^11.36.0"
tika = "^2.6.0"
ofxparse = "^0.21"
numpy = "^1.24.1"
python-dateutil = "^2.8.2"
alembic = "^1.9.1"
sqlalchemy = {extras = ["mypy"], version = "^2.0.4"}
pydantic = {extras = ["email"], version = "^1.10.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.5"
emails = "^0.6"
python-multipart = "^0.0.6"
celery = "^5.2.7"
tenacity = "^8.1.0"
httpx = "^0.23.3"
raven = "^6.10.0"
jinja2 = "^3.1.2"
lau-commons = {git = "https://github.com/renatocarnauba/lau_commons.git"}
bson = "^0.5.10"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
mypy = "^1.0.1"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.20.3"
anyio = {extras = ["trio"], version = "^3.6.2"}
pytest-pretty = "^1.1.0"
autoflake = "^2.0.0"
flake8 = "^6.0.0"
autopep8 = "^2.0.1"
types-python-jose = "^3.3.4.4"
types-passlib = "^1.7.7.9"
faker = "^17.0.0"
flake8-pyproject = "^1.2.2"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.flake8]
max-line-length = 120
exclude = [
  ".git",
  "__pycache__",
  "__init__.py",
  ".mypy_cache",
  ".pytest_cache",
  "./alembic/",
  "./_deprecated/"
]


[tool.mypy]
plugins = [
  "pydantic.mypy",
  "sqlalchemy.ext.mypy.plugin"
]
exclude = "app/tests"
ignore_missing_imports = true
check_untyped_defs = true
disallow_untyped_defs = true
follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true